// src/deltas/domain/enums.ts
/**
 * Auto generated. DO NOT edit manually.
 * Last updated on: Fri, 14 Jul 2023 19:19:52 GMT
 */

/**
 * @enum
 * @readonly
 * @description RefWorlds Table Names
 */
export enum RefWorldsTableName {
    Config = 'config',
    Deposits = 'deposits',
    Referendums = 'referendums',
    Votes = 'votes',
};